Adds own file for theme values
