+++ src/components/Theme/index.jsx
@@ -1,2 +1,2 @@
-export { default as Theme } from "./Theme";
+export { default as Theme } from "./ThemeWrapper";
 export { default as GlobalStyle } from "./GlobalStyle";
