+++ src/contexts/ThemeContext.jsx
@@ -1,46 +1,7 @@
 import { createContext } from "react";
 import useLocalStorage from "../hooks/useLocalStorage";
 import { adjustThemeColors } from "../utils/";
-
-const theme = {
-  colors: {
-    primary: "hsl(0, 78.87323943661971%, 72.15686274509804%)",
-    secondary: "hsl(152.54237288135593, 100%, 65.29411764705883%)",
-    tertiary: "hsl(15.862068965517242, 100%, 65.88235294117646%)",
-    attention: "hsl(60, 100%, 50.78431372549019%)",
-    warning: "hsl(0, 100%, 65.29411764705883%)",
-    success: "hsl(152.54237288135593, 100%, 65.29411764705883%)",
-    active: "hsl(0, 100%, 65.29411764705883%)",
-    black: "hsl(0, 0%, 0%)",
-    white: "hsl(0, 0%, 100%)",
-    disabled: "hsl(0, 0%, 82.35294117647058%)",
-    grey: "hsl(0, 0%, 34.90196078431372%)",
-    background: "hsl(0, 78.87323943661971%, 72.15686274509804%)",
-  },
-  borders: {
-    regular: "2px solid",
-  },
-  shadows: {
-    small: "5px 5px 0",
-    medium: "8px 8px 0",
-    regular: "10px 10px 0",
-    large: "15px 15px 0",
-    xl: "20px 20px 0",
-  },
-  translate: {
-    small: "-5px -5px",
-    regular: "-10px -10px",
-  },
-  borderRadius: {
-    small: "10px",
-    regular: "15px",
-    image: "13px",
-  },
-  fonts: {
-    heading: "IBM Plex Mono, Arial, Helvetica, sans-serif",
-    body: "Inter, Arial, Helvetica, sans-serif",
-  },
-};
+import theme from "../components/Theme/theme";
 
 export const ThemeContext = createContext();
